Move TMem tests to a separate file (#4119)
